pyproject.toml
[tool.poetry]
name = "mmv"
version = "4.1.2"
description = "The Interactive Shader Render Platform"
authors = ["Tremeschin"]
license = "GPLv3"
packages = [
    {include = "App"}
]

[tool.poetry.dependencies]
python = ">=3.7,<3.10"
numpy = "^1.20.3"
Pillow = "^8.2.0"
moderngl = "^5.6.4"
moderngl-window = "^2.3.0"
PyYAML = "^5.4.1"
mido = "^1.2.10"
tqdm = "^4.61.1"
opencv-python-headless = "^4.5.2"
wheel = "^0.36.2"
imgui = "^1.3.0"
typer = "^0.3.2"
watchdog = "^2.1.2"
scipy = "^1.6.3"
ujson = "^4.0.2"
pyFluidSynth = "^1.3.0"
dotmap = "^1.3.23"
numpy-quaternion = "^2021.6.9"
intervaltree = "^3.1.0"
wonderwords = "^2.2.0"
pygame = "^2.0.1"
playsound = "^1.2.2"
toml = "^0.10.2"
requests = "^2.25.1"
arrow = "^1.1.1"
SoundCard = "^0.4.1"
glfw = "^2.1.0"
appdirs = "^1.4.4"
dearpygui = "0.8.39"

[tool.poetry.dev-dependencies]
Nuitka = "^0.6.16"
scons = "^4.1.0"
imageio = "^2.9.0"
zstandard = "^0.15.2"

[tool.poetry.scripts]
magic = "App.Magic:Main"
release = "App.MakeRelease:Main"
editor = "App.RunEditor:Main"
prototype = "App.Prototype:Main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "testpypi"
url = "https://test.pypi.org/simple"
